refactor(types): move Parsed and the storage types into leaf packages
Groundwork for making pkg/config importable from a server (#48). Both types
are plain data living in packages made heavy by their functions, which is
the same shape pkg/urlprotocol addressed in #47.

repourl.Parsed is Protocol plus three strings, but pkg/repository/url costs
43 packages because url.go reaches oops, samber/lo and golang.org/x/text.
Moved to pkg/urlprotocol alongside Protocol, taking AsHTTPsURL and HostName
with it — both stdlib-only, and methods cannot stay behind on an aliased
type. Parse stays where it is.

storageplan.StoragePlan and Disk are equally plain, but the package costs
142 packages via its executor, shell templates and lipgloss tree printing.
disk.go imported only pkg/constants, so it moved whole; StoragePlan's three
methods could not follow, so getUITree, PrettyPrint and Execute became
functions — three call sites.

Both original packages alias the moved types and re-export NewDisk, so
nothing else changes.
